Move plotting function to vis_utils
